uint40 for timestamps, uint72 for feed IDs
This is to help prevent accidentally confusing values that have the
same type with each other.  (This is particularly error-prone when
calling functions that take many parameters or return many values.)

Since uint72 is a very uncommon type, client code benefits by not
being able to accidentally mix up a feed ID with something else.

Same goes for the timestamp, for which uint40 is the smallest type
that can comfortably accomodate thousands of years into the future.